LayoutTests/imported/w3c/web-platform-tests/css/css-values/lh-unit-003-expected.txt

X
X
X
X
X
X
X
X
X
X

PASS Line-height and lh before @font-face loads
FAIL Line-height and lh after @font-face loaded assert_equals: lh and normal line-height both affected expected 269 but got 260
